REF: hd wallet classes
