+++ class/hd-segwit-p2sh-wallet.js
@@ -50,32 +50,6 @@ export class HDSegwitP2SHWallet extends AbstractHDWallet {
     });
   }
 
-  async fetchBalance() {
-    try {
-      const api = new Frisbee({ baseURI: 'https://www.blockonomics.co' });
-      let response = await api.post('/api/balance', { body: JSON.stringify({ addr: this.getXpub() }) });
-      // console.log(response);
-
-      if (response && response.body && response.body.response) {
-        this.balance = 0;
-        this.unconfirmed_balance = 0;
-        this.usedAddresses = [];
-        for (let addr of response.body.response) {
-          this.balance += addr.confirmed;
-          this.unconfirmed_balance += addr.unconfirmed;
-          this.usedAddresses.push(addr.addr);
-        }
-        this.balance = new BigNumber(this.balance).dividedBy(100000000).toString() * 1;
-        this.unconfirmed_balance = new BigNumber(this.unconfirmed_balance).dividedBy(100000000).toString() * 1;
-        this._lastBalanceFetch = +new Date();
-      } else {
-        throw new Error('Could not fetch balance from API: ' + response.err);
-      }
-    } catch (err) {
-      console.warn(err);
-    }
-  }
-
   _getExternalWIFByIndex(index) {
     index = index * 1; // cast to int
     let mnemonic = this.secret;
@@ -288,61 +262,6 @@ export class HDSegwitP2SHWallet extends AbstractHDWallet {
     } catch (err) {
       console.warn(err);
     }
-  }
-
-  /**
-   * @inheritDoc
-   */
-  async fetchUtxo() {
-    const api = new Frisbee({
-      baseURI: 'https://blockchain.info',
-    });
-
-    if (this.usedAddresses.length === 0) {
-      // just for any case, refresh balance (it refreshes internal `this.usedAddresses`)
-      await this.fetchBalance();
-    }
-
-    let addresses = this.usedAddresses.join('|');
-    addresses += '|' + this._getExternalAddressByIndex(this.next_free_address_index);
-    addresses += '|' + this._getInternalAddressByIndex(this.next_free_change_address_index);
-
-    let utxos = [];
-
-    let response;
-    try {
-      response = await api.get('/unspent?active=' + addresses + '&limit=1000');
-      // this endpoint does not support offset of some kind o_O
-      // so doing only one call
-      let json = response.body;
-      if (typeof json === 'undefined' || typeof json.unspent_outputs === 'undefined') {
-        throw new Error('Could not fetch UTXO from API ' + response.err);
-      }
-
-      for (let unspent of json.unspent_outputs) {
-        // a lil transform for signer module
-        unspent.txid = unspent.tx_hash_big_endian;
-        unspent.vout = unspent.tx_output_n;
-        unspent.amount = unspent.value;
-
-        let chunksIn = bitcoin.script.decompile(Buffer.from(unspent.script, 'hex'));
-        unspent.address = bitcoin.address.fromOutputScript(chunksIn);
-        utxos.push(unspent);
-      }
-    } catch (err) {
-      console.warn(err);
-    }
-
-    this.utxo = utxos;
-  }
-
-  weOwnAddress(addr) {
-    let hashmap = {};
-    for (let a of this.usedAddresses) {
-      hashmap[a] = 1;
-    }
-
-    return hashmap[addr] === 1;
   }
 
   createTx(utxos, amount, fee, address) {
